Reduction of compiler warnings
Many tweaks mostly to silence compiler warnings and remove some dead code and variables.

Jira-id: ESSBM-460

+++ driver-examples/mss/mss-ethernet-mac/mpfs-uart-mac-freertos_lwip/src/platform/drivers/mss/mss_ethernet_mac/mss_ethernet_mac.c
@@ -5484,7 +5484,7 @@ rxpkt_handler(mss_mac_instance_t *this_mac, uint64_t queue_no)
 
                 this_queue->pckt_rx_callback(
                     this_mac,
-                    queue_no,
+                    (uint32_t)queue_no,
                     p_rx_packet,
                     pckt_length,
                     cdesc,
@@ -5562,7 +5562,7 @@ txpkt_handler(mss_mac_instance_t *this_mac, uint64_t queue_no)
                 {
                     this_queue->pckt_tx_callback(
                         this_mac,
-                        queue_no,
+                        (uint32_t)queue_no,
                         p_current_desc,
                         this_queue->tx_caller_info[this_queue->current_tx_desc]);
                 }
